// src/app/utils/interfaces/normativa.ts
export interface Normativa {
    procesoEvaluacionId: number
    indicadorSiacid: number
    componenteId: number
    elementoEvaluacionId: number
    areaCentralId: number
    subAreaCentralId: number
    normativaId: number
    claveNormatividad: string
    normatividad: string
    activo: boolean
    usuarioCreacion: number
    usuarioModificacion: number
}